\name{AHMeSHDbs-package}
\alias{AHMeSHDbs-package}
\alias{AHMeSHDbs}
\docType{package}
\title{
\packageTitle{AHMeSHDbs}
}
\description{
\packageDescription{AHMeSHDbs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AHMeSHDbs}
\packageIndices{AHMeSHDbs}
}
\author{
\packageAuthor{AHMeSHDbs}

Maintainer: \packageMaintainer{AHMeSHDbs}
}

\keyword{ package }
\examples{
ls("package:AHMeSHDbs") # This is empty package
}